layout/generic/test/test_bug673315-1.html

<!DOCTYPE>
<html>
<head>
<title>test for bug 673315 part.1</title>
<script type="text/javascript"
  src="chrome://mochikit/content/MochiKit/packed.js"></script>
<script type="text/javascript"
  src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<script type="text/javascript"
  src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
<link rel="stylesheet" type="text/css"
  href="chrome://mochikit/content/tests/SimpleTest/test.css" />
</head>
<body>
<p id="p">aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa</p>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();

var gSelectionListener = {
  reason: 0,
  reset: function () { this.reason = 0; },
  notifySelectionChanged: function (doc, sel, reason)
  {
    this.reason |= reason;
  }
}

function doTest() {
  var p = document.getElementById("p");

  const nsISelectionPrivate = Components.interfaces.nsISelectionPrivate;
  const nsISelectionListener = Components.interfaces.nsISelectionListener;
  var selPrivate =
    window.getSelection().QueryInterface(nsISelectionPrivate);
  selPrivate.addSelectionListener(gSelectionListener);

  gSelectionListener.reset();
  synthesizeMouse(p, 10, 10, { type: "mousedown", button: 0 });
  is(gSelectionListener.reason,
     nsISelectionListener.DRAG_REASON | nsISelectionListener.MOUSEDOWN_REASON,
     "the reason isn't drag and mosuedown after mousedown event");
  gSelectionListener.reset();
  synthesizeMouse(p, 10, 50, { type: "mousemove" });
  is(gSelectionListener.reason,
     nsISelectionListener.DRAG_REASON | nsISelectionListener.MOUSEDOWN_REASON,
     "the reason isn't drag and mosuedown after mousemove event");
  gSelectionListener.reset();
  synthesizeMouse(p, 10, 50, { type: "mouseup", button: 0 });
  is(gSelectionListener.reason,
     nsISelectionListener.MOUSEUP_REASON,
     "the reason isn't mosueup after mouseup event");

  selPrivate.addSelectionListener(gSelectionListener);

  SimpleTest.finish();
}

SimpleTest.waitForFocus(doTest, window);

</script>
</pre>
</body>
</html>
